Cleaning up newfreq
Looking at this file, the class is pretty messy, plus some of the routines and checks don't really need to be there. I'm not sure about the best way to handle things... I'm thinking a handful of flags for what I'm checking would be easiest and most transparent, so this will need some work.