Only verify isotope if in right format
Split the isotope saving function into two, just so we don't try to check if the isotope is in the table if we know that we don't have the right format. The leading dash (if the user tries to put the isotope in the form ###-SYM or similar) is clipped off. If they put two dashes, they deserve to have the program tell them not to.