Tabular output implemented
I did some tweaking in the Isotope class to make the header for the calculations look better as well. Most of the work was in Display, but it now shows the energy/frequency pairs with blanks where the energy would've fallen out of range of the desired one (unlike previously that just extended the range to include more values so that it was a completely full grid).